Extract some reusable test components and split some test files
