+++ test/test_helper_nets.h
@@ -8,4 +8,6 @@ void set_majority_net_multi_cycle_embedded(int *majority, int node_ct);
 void set_majority_net_two_embedded_cycles(int *majority, int node_ct);
 void set_majority_tarjan_example(int *majority, int node_ct);
 
+void set_preference_net_1(int *preference_graph, int node_ct);
+
 #endif
